+++ xmldirector/plonecore/browser/__init__.py
@@ -4,3 +4,5 @@
 # xmldirector.plonecore
 # (C) 2014,  Andreas Jung, www.zopyx.com, Tuebingen, Germany
 ################################################################
+
+import restapi_patches